refactor(backend): move find_recent_species logic from database to species service
